+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Support for Nextcloud 19
 - Argon2id support
 - System wide values option
+- Allow email login option
 
 ## [4.4.1] - 2020-02-02
 ### Fixed